Add missing directory creation for railsweb node
I think we modified the volume configuration to put the redis persistence on the railsweb node while working through the deployment of dashboard.wikiedu.org production, but we must have manually created the data directory.

Without that redis directory, nomad can't actually run this node as a client, it doesn't show up on the topology UI, and it can't place the jobs that require the rails or redis volumes.