Extend nginx config to handle Nomad UI
I successfully added a reverse proxy for the Nomad UI to a test cluster today. These same steps should work to update the nginx config on the dashboard.wikiedu.org production cluster, by first adding the port 80 server and setting up the SSL cert, and then updating to this commit to deploy the reverse proxy.