Refactor terragrunt config to prepare for multi-site config
This adds a staging environment directory, and moves the terragrunt config files to it. In this setup, the terraform state all lives in the particular environment directory, and each environment has its own secrets files. I haven't tested moving the state file and confirming that existing infrastructure can survive this refactor, but it looks like it should work.