Update README for multiple environments, extend deletion script
For some reason, the part of the setup where it writes to `nomad.sh` isn't working now when there are two environments. However, during initial `apply`, it does still write that nomad.sh file to the terragrunt cache, so we can work around it by moving it to the expected location manually.