Update nginx renewal script and crontab instructions
I think I finally worked out the remaining problems with automatically renewing the certbot certs! The script previously worked when run manually (via `bash certbot_renew.sh`) but was failing via cron. This was because of the `-it` flags in the docker commands, which would fail when run outside of a terminal environment.